Refactor import statements in plctranslator package
