+++ src/plctranslator/__init__.py
@@ -1,4 +1,5 @@
 """Package for PLC Translator."""
 
 # ruff: noqa: F401 reason: This is a package.
-from .tia_translator import generate_code, generate_variable_text, read_scl_file
+from .tia_helpers import read_scl_file
+from .tia_translator import generate_code, generate_variable_text
